+++ CameraOrtho.hpp
@@ -8,11 +8,11 @@
 
 #include "ICamera.hpp"
 
-namespace mygl 
+namespace mygl
 {
     class CameraOrtho : public ICamera {
         public:
-            CameraOrtho(glm::vec3 position = glm::vec3(0.0f, 0.0f, 0.0f), float win_width = 800, float win_height = 600);
+            CameraOrtho(glm::vec3 position = glm::vec3(0.0f, 0.0f, 0.0f), int win_width = 800, int win_height = 600);
             glm::mat4 getViewMatrix() const;
             glm::mat4 getProjectionMatrix() const;
             void processMouseScroll(float yoffset);
@@ -23,10 +23,10 @@ namespace mygl
             glm::vec3 up;
             glm::vec3 right;
             glm::vec3 world_up;
-            
+
             float yaw;
             float pitch;
-            
+
             float movement_speed;
             float mouse_sensitivity;
             float fov;
